Better layout for easy mode, better handling of overlays
