--- www/src/Pages/Leds/AnimationSection.scss
@@ -1,22 +1,0 @@
-.case-grid {
-	display: grid;
-	grid-template-columns: repeat(5, 1fr);
-}
-
-@media only screen and (max-width: 800px) {
-	.case-grid {
-		grid-template-columns: repeat(4, 1fr);
-	}
-}
-@media only screen and (max-width: 680px) {
-	.case-grid {
-		grid-template-columns: repeat(2, 1fr);
-	}
-}
-
-@media only screen and (max-width: 500px) {
-	.case-grid {
-		display: flex;
-		flex-direction: column;
-	}
-}
