+++ www/src/Pages/Leds/LightIndicator.tsx
@@ -1,6 +1,6 @@
 import { useDraggable } from '@dnd-kit/core';
-import { OverlayTrigger, Tooltip } from 'react-bootstrap';
-import { Light } from '../Store/useLedStore';
+import { OverlayTrigger, Popover } from 'react-bootstrap';
+import { Light } from '../../Store/useLedStore';
 
 type LightIndicatorProps = {
 	id: number;
@@ -9,14 +9,22 @@ type LightIndicatorProps = {
 	error?: string;
 } & Light;
 
-export const LightIcon = ({ size, active }: { size: number; active: boolean }) => (
+export const LightIcon = ({
+	size,
+	active,
+	error,
+}: {
+	size: number;
+	active: boolean;
+	error?: string;
+}) => (
 	<div>
 		{active ? (
 			<svg
 				xmlns="http://www.w3.org/2000/svg"
 				width={size}
 				height={size}
-				fill="currentColor"
+				fill={error ? 'red' : 'currentColor'}
 				viewBox="0 0 16 16"
 				style={{ filter: 'drop-shadow(0 4px 16px rgba(255, 255, 200, 0.7))' }}
 			>
@@ -27,7 +35,7 @@ export const LightIcon = ({ size, active }: { size: number; active: boolean }) =
 				xmlns="http://www.w3.org/2000/svg"
 				width={size}
 				height={size}
-				fill="currentColor"
+				fill={error ? 'red' : 'currentColor'}
 				viewBox="0 0 16 16"
 				style={{ filter: 'drop-shadow(0 4px 16px rgba(255, 255, 200, 0.7))' }}
 			>
@@ -77,19 +85,17 @@ export function LightIndicator({
 			}}
 		>
 			{isDragging ? (
-				<LightIcon size={cellWidth} active={active} />
+				<LightIcon size={cellWidth} active={active} error={error} />
 			) : (
 				<OverlayTrigger
-					placement="top"
-					{...(error ? { show: true } : {})}
+					key={`light-${id}`}
+					placement="auto"
+					rootClose
 					overlay={
-						<Tooltip>
-							{error ? (
-								<div>{error}</div>
-							) : (
-								<div className="d-flex flex-column align-items-start min-w-110 p-1">
-									<div className="fw-semibold mb-1">Light {id + 1}</div>
-									<hr className="m-1 w-100" style={{ margin: '1px 0' }} />
+						<Popover>
+							<Popover.Header as="h3">Light {id + 1}</Popover.Header>
+							<Popover.Body>
+								<div style={{ minWidth: 200 }}>
 									<div className="d-flex w-100 justify-content-between">
 										<span className="text-secondary">GPIO/Case:</span>
 										<span>{GPIOPinorCaseChainIndex}</span>
@@ -118,12 +124,12 @@ export function LightIndicator({
 										<span>{firstLedIndex}</span>
 									</div>
 								</div>
-							)}
-						</Tooltip>
+							</Popover.Body>
+						</Popover>
 					}
 				>
 					<div>
-						<LightIcon size={cellWidth} active={active} />
+						<LightIcon size={cellWidth} active={active} error={error} />
 					</div>
 				</OverlayTrigger>
 			)}
